requirements.txt
a2wsgi==1.10.8
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiologic==0.14.0
aiosignal==1.3.2
aiosqlite==0.21.0
airflow-operators==0.7.0
alembic==1.16.1
annotated-types==0.7.0
anthropic==0.54.0
anyio==4.9.0
apache-airflow==3.0.2
apache-airflow-core==3.0.2
apache-airflow-providers-apache-kafka==1.9.0
apache-airflow-providers-common-compat==1.7.0
apache-airflow-providers-common-io==1.6.0
apache-airflow-providers-common-sql==1.27.1
apache-airflow-providers-mysql==5.7.4
apache-airflow-providers-qdrant==1.4.0
apache-airflow-providers-smtp==2.1.0
apache-airflow-providers-standard==1.2.0
apache-airflow-task-sdk==1.0.2
apispec==6.8.2
appnope==0.1.4
argcomplete==3.6.2
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
asyncer==0.0.7
asyncpg==0.30.0
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
banks==2.1.2
beautifulsoup4==4.13.4
bidict==0.23.1
bleach==6.2.0
blinker==1.9.0
boto3==1.38.36
botocore==1.38.36
cadwyn==5.4.2
certifi==2025.6.15
cffi==1.17.1
chainlit==2.5.5
charset-normalizer==3.4.2
chevron==0.14.0
click==8.2.1
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
confluent-kafka==2.10.1
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==45.0.4
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
deprecated==1.2.18
dill==0.4.0
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
email-validator==2.2.0
executing==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
fastjsonschema==2.21.1
filelock==3.18.0
filetype==1.2.0
flask==2.3.3
flask-appbuilder==4.7.0
flask-babel==2.0.0
flask-jwt-extended==4.7.1
flask-limiter==3.12
flask-login==0.6.3
flask-sqlalchemy==2.5.1
flask-wtf==1.2.2
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.5.1
googleapis-common-protos==1.70.0
greenlet==3.2.3
griffe==1.7.3
grpcio==1.73.0
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.3
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
hyperframe==6.1.0
idna==3.10
importlib-metadata==8.6.1
inflection==0.5.1
ipykernel==6.29.5
ipython==9.3.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.8.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server==2.16.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.3
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.15
lazify==0.4.0
lazy-object-proxy==1.11.0
libcst==1.8.1
limits==5.3.0
linkify-it-py==2.0.3
literalai==0.1.201
llama-cloud==0.1.26
llama-cloud-services==0.6.32
llama-index==0.12.42
llama-index-agent-openai==0.4.11
llama-index-cli==0.4.3
llama-index-core==0.12.42
llama-index-embeddings-ollama==0.6.0
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.7.7
llama-index-llms-ollama==0.6.2
llama-index-llms-openai==0.4.7
llama-index-multi-modal-llms-openai==0.5.1
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.9
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-qdrant==0.6.0
llama-parse==0.6.32
lockfile==0.12.2
mako==1.3.10
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.26.1
marshmallow-sqlalchemy==1.4.2
matplotlib-inline==0.1.7
mcp==1.9.4
mdurl==0.1.2
methodtools==0.4.7
mistune==3.1.3
monotonic==1.6
more-itertools==10.7.0
msgspec==0.19.0
multidict==6.4.4
mypy-extensions==1.1.0
mysql-connector-python==9.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
notebook==7.4.3
notebook-shim==0.2.4
numpy==2.3.0
ollama==0.5.1
openai==1.86.0
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-alephalpha==0.40.9
opentelemetry-instrumentation-anthropic==0.40.9
opentelemetry-instrumentation-bedrock==0.40.9
opentelemetry-instrumentation-chromadb==0.40.9
opentelemetry-instrumentation-cohere==0.40.9
opentelemetry-instrumentation-crewai==0.40.9
opentelemetry-instrumentation-google-generativeai==0.40.9
opentelemetry-instrumentation-groq==0.40.9
opentelemetry-instrumentation-haystack==0.40.9
opentelemetry-instrumentation-lancedb==0.40.9
opentelemetry-instrumentation-langchain==0.40.9
opentelemetry-instrumentation-llamaindex==0.40.9
opentelemetry-instrumentation-logging==0.52b1
opentelemetry-instrumentation-marqo==0.40.9
opentelemetry-instrumentation-mcp==0.40.9
opentelemetry-instrumentation-milvus==0.40.9
opentelemetry-instrumentation-mistralai==0.40.9
opentelemetry-instrumentation-ollama==0.40.9
opentelemetry-instrumentation-openai==0.40.9
opentelemetry-instrumentation-pinecone==0.40.9
opentelemetry-instrumentation-qdrant==0.40.9
opentelemetry-instrumentation-replicate==0.40.9
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-sagemaker==0.40.9
opentelemetry-instrumentation-sqlalchemy==0.52b1
opentelemetry-instrumentation-threading==0.52b1
opentelemetry-instrumentation-together==0.40.9
opentelemetry-instrumentation-transformers==0.40.9
opentelemetry-instrumentation-urllib3==0.52b1
opentelemetry-instrumentation-vertexai==0.40.9
opentelemetry-instrumentation-watsonx==0.40.9
opentelemetry-instrumentation-weaviate==0.40.9
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-semantic-conventions-ai==0.4.9
opentelemetry-util-http==0.52b1
ordered-set==4.1.0
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pendulum==3.1.0
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
portalocker==2.10.1
posthog==3.25.0
prison==0.2.1
prometheus-client==0.22.1
prompt-toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.9.1
pygments==2.19.1
pyjwt==2.10.1
pypdf==5.6.0
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.2
python-json-logger==3.3.0
python-multipart==0.0.18
python-slugify==8.0.4
python-socketio==5.13.0
pytz==2025.2
pyyaml==6.0.2
pyzmq==27.0.0
qdrant-client==1.14.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
retryhttp==1.3.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-argparse==1.7.1
rich-toolkit==0.14.7
rpds-py==0.25.1
s3transfer==0.13.0
send2trash==1.8.3
setproctitle==1.3.6
setuptools==80.9.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==1.4.54
sqlalchemy-jsonfield==1.0.2
sqlalchemy-utils==0.41.2
sqlparse==0.5.3
sse-starlette==2.3.6
stack-data==0.6.3
starlette==0.41.3
striprtf==0.0.26
structlog==25.4.0
svcs==25.1.0
syncer==2.0.3
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
terminado==0.18.1
text-unidecode==1.3
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
tomli==2.2.1
tornado==6.5.1
tqdm==4.67.1
traceloop-sdk==0.40.9
traitlets==5.14.3
typer==0.16.0
types-python-dateutil==2.9.0.20250516
types-requests==2.32.4.20250611
typing-extensions==4.13.2
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
uc-micro-py==1.0.3
universal-pathlib==0.2.6
uptrace==1.31.0
uri-template==1.3.0
urllib3==2.4.0
uuid6==2025.0.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==0.20.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
werkzeug==3.1.3
widgetsnbextension==4.0.14
wirerope==1.0.0
wrapt==1.17.2
wsproto==1.2.0
wtforms==3.2.1
yarl==1.20.1
zipp==3.23.0
